+++ rye/src/cli/add.rs
@@ -32,7 +32,7 @@ if py_ver:
     finder.target_python.py_ver = tuple(map(int, py_ver.split('.')))
 choices = iter(finder.find_matches(package))
 if not pre:
-    choices = (m for m in choices if not Version(m.version).is_prerelease)
+    choices = (m for m in choices if not(m.version and Version(m.version).is_prerelease))
 
 print(json.dumps([x.as_json() for x in choices]))
 "#;
@@ -40,7 +40,7 @@ print(json.dumps([x.as_json() for x in choices]))
 #[derive(Deserialize, Debug)]
 struct Match {
     name: String,
-    version: String,
+    version: Option<String>,
     link: Option<Link>,
 }
 
@@ -220,7 +220,7 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
                             eprintln!(
                                 "  {} ({}) requires Python {}",
                                 pkg.name,
-                                pkg.version,
+                                pkg.version.unwrap_or_default(),
                                 pkg.link
                                     .as_ref()
                                     .and_then(|x| x.requires_python.as_ref())
@@ -237,9 +237,9 @@ pub fn execute(cmd: Args) -> Result<(), Error> {
             }
 
             let m = matches.into_iter().next().unwrap();
-            if requirement.version_or_url.is_none() {
+            if m.version.is_some() && requirement.version_or_url.is_none() {
                 requirement.version_or_url = Some(VersionOrUrl::VersionSpecifier(
-                    VersionSpecifiers::from_str(&format!("~={}", m.version))?,
+                    VersionSpecifiers::from_str(&format!("~={}", m.version.unwrap()))?,
                 ));
             }
             requirement.name = m.name;