+++ rye/src/cli/init.rs
@@ -84,6 +84,11 @@ build-backend = "filt_core.buildapi"
 [tool.rye]
 managed = true
 
+{%- if build_system == "hatchling" %}
+[tool.hatch.metadata]
+allow-direct-references = true
+{%- endif %}
+
 "#;
 
 /// The template for the readme file.